Move and update offers utils
